refactor: update icon picker component and styles for improved flexibility
- Removed the `rootStyle` prop from the IconPicker component to streamline styling.
- Enhanced the `useStyles` function to support optional styling models, allowing for more dynamic styling configurations.
- Updated the IconPickerWrapper to eliminate unnecessary style props, ensuring cleaner integration with the icon picker.
- Removed the deprecated styles file, consolidating styles into a single location for better maintainability.
- Adjusted settings form to exclude additional font properties for a more focused configuration.
